requirements.txt
configargparse==1.0
gensim==3.8.1
numpy==1.17.0
pandas==0.25.3
pytorch==1.3.0
scikit-learn==0.22.1
